[user-service] extract user-authentication
