+++ components/server/src/api/user.spec.ts
@@ -9,7 +9,7 @@ import { APIUserService } from "./user";
 import { Container } from "inversify";
 import { testContainer } from "@gitpod/gitpod-db/lib";
 import { WorkspaceStarter } from "../workspace/workspace-starter";
-import { UserService } from "../user/user-service";
+import { UserAuthentication } from "../user/user-authentication";
 import { BlockUserRequest, BlockUserResponse } from "@gitpod/public-api/lib/gitpod/experimental/v1/user_pb";
 import { User } from "@gitpod/gitpod-protocol";
 import { StopWorkspacePolicy } from "@gitpod/ws-manager/lib";
@@ -34,19 +34,19 @@ export class APIUserServiceSpec {
             return [];
         },
     } as WorkspaceStarter;
-    private userServiceMock: UserService = {
+    private userServiceMock: UserAuthentication = {
         blockUser: async (targetUserId: string, block: boolean): Promise<User> => {
             return {
                 id: targetUserId,
             } as User;
         },
-    } as UserService;
+    } as UserAuthentication;
 
     async before() {
         this.container = testContainer.createChild();
 
         this.container.bind(WorkspaceStarter).toConstantValue(this.workspaceStarterMock);
-        this.container.bind(UserService).toConstantValue(this.userServiceMock);
+        this.container.bind(UserAuthentication).toConstantValue(this.userServiceMock);
         this.container.bind(APIUserService).toSelf().inSingletonScope();
     }
