+++ components/server/src/user/user-service.spec.db.ts
@@ -8,11 +8,14 @@ import { Experiments } from "@gitpod/gitpod-protocol/lib/experiments/configcat-s
 import * as chai from "chai";
 import { Container } from "inversify";
 import { createTestContainer } from "../test/service-testing-container-module";
-import { UserService } from "./user-service";
 import { BUILTIN_INSTLLATION_ADMIN_USER_ID, TypeORM } from "@gitpod/gitpod-db/lib";
 import { resetDB } from "@gitpod/gitpod-db/lib/test/reset-db";
 import { OrganizationService } from "../orgs/organization-service";
 import { Authorizer } from "../authorization/authorizer";
+import { UserService } from "./user-service";
+import { Organization, User } from "@gitpod/gitpod-protocol";
+import { ErrorCodes } from "@gitpod/gitpod-protocol/lib/messaging/error";
+import { expectError } from "../test/expect-utils";
 
 const expect = chai.expect;
 
@@ -20,6 +23,11 @@ describe("UserService", async () => {
     let container: Container;
     let userService: UserService;
     let auth: Authorizer;
+    let org: Organization;
+    let user: User;
+    let user2: User;
+    let nonOrgUser: User;
+    let admin: User;
 
     beforeEach(async () => {
         container = createTestContainer();
@@ -28,6 +36,38 @@ describe("UserService", async () => {
         });
         userService = container.get<UserService>(UserService);
         auth = container.get(Authorizer);
+        const orgService = container.get<OrganizationService>(OrganizationService);
+        org = await orgService.createOrganization(BUILTIN_INSTLLATION_ADMIN_USER_ID, "myOrg");
+        user = await userService.createUser({
+            organizationId: org.id,
+            identity: {
+                authId: "foo",
+                authName: "bar",
+                authProviderId: "github",
+                primaryEmail: "yolo@yolo.com",
+            },
+        });
+
+        user2 = await userService.createUser({
+            organizationId: org.id,
+            identity: {
+                authId: "foo",
+                authName: "bar",
+                authProviderId: "github",
+                primaryEmail: "yolo@yolo.com",
+            },
+        });
+
+        nonOrgUser = await userService.createUser({
+            identity: {
+                authId: "foo",
+                authName: "bar",
+                authProviderId: "github",
+                primaryEmail: "yolo@yolo.com",
+            },
+        });
+
+        admin = await userService.findUserById(BUILTIN_INSTLLATION_ADMIN_USER_ID, BUILTIN_INSTLLATION_ADMIN_USER_ID);
     });
 
     afterEach(async () => {
@@ -36,44 +76,9 @@ describe("UserService", async () => {
     });
 
     it("createUser", async () => {
-        const orgService = container.get<OrganizationService>(OrganizationService);
-        const org = await orgService.createOrganization(BUILTIN_INSTLLATION_ADMIN_USER_ID, "myOrg");
-        const user = await userService.createUser({
-            organizationId: org.id,
-            identity: {
-                authId: "foo",
-                authName: "bar",
-                authProviderId: "github",
-                primaryEmail: "yolo@yolo.com",
-            },
-        });
-
-        const user2 = await userService.createUser({
-            organizationId: org.id,
-            identity: {
-                authId: "foo",
-                authName: "bar",
-                authProviderId: "github",
-                primaryEmail: "yolo@yolo.com",
-            },
-        });
-
-        const nonOrgUser = await userService.createUser({
-            identity: {
-                authId: "foo",
-                authName: "bar",
-                authProviderId: "github",
-                primaryEmail: "yolo@yolo.com",
-            },
-        });
-
         expect(await auth.hasPermissionOnUser(user.id, "read_info", user.id)).to.be.true;
         expect(await auth.hasPermissionOnUser(user.id, "write_info", user.id)).to.be.true;
         expect(await auth.hasPermissionOnUser(user.id, "suspend", user.id)).to.be.true;
-
-        expect(await auth.hasPermissionOnUser(BUILTIN_INSTLLATION_ADMIN_USER_ID, "read_info", user.id)).to.be.true;
-        expect(await auth.hasPermissionOnUser(BUILTIN_INSTLLATION_ADMIN_USER_ID, "write_info", user.id)).to.be.true;
-        expect(await auth.hasPermissionOnUser(BUILTIN_INSTLLATION_ADMIN_USER_ID, "suspend", user.id)).to.be.true;
 
         expect(await auth.hasPermissionOnUser(user2.id, "read_info", user.id)).to.be.true;
         expect(await auth.hasPermissionOnUser(user2.id, "write_info", user.id)).to.be.false;
@@ -82,23 +87,26 @@ describe("UserService", async () => {
         expect(await auth.hasPermissionOnUser(nonOrgUser.id, "read_info", user.id)).to.be.false;
         expect(await auth.hasPermissionOnUser(nonOrgUser.id, "write_info", user.id)).to.be.false;
         expect(await auth.hasPermissionOnUser(nonOrgUser.id, "suspend", user.id)).to.be.false;
+
+        expect(await auth.hasPermissionOnUser(admin.id, "read_info", user.id)).to.be.true;
+        expect(await auth.hasPermissionOnUser(admin.id, "write_info", user.id)).to.be.true;
+        expect(await auth.hasPermissionOnUser(admin.id, "suspend", user.id)).to.be.true;
     });
 
     it("updateLoggedInUser_avatarUrlNotUpdatable", async () => {
-        const user = await userService.createUser({
-            identity: {
-                authId: "foo",
-                authName: "bar",
-                authProviderId: "github",
-                primaryEmail: "yolo@yolo.com",
-            },
-        });
-
-        const updated = await userService.updateUser(user.id, {
+        const update = {
+            id: user.id,
             avatarUrl: "evil-payload",
-        });
+            additionalData: {
+                disabledClosedTimeout: true,
+            },
+        };
+        await expectError(ErrorCodes.NOT_FOUND, userService.updateUser(nonOrgUser.id, update));
+        await expectError(ErrorCodes.PERMISSION_DENIED, userService.updateUser(user2.id, update));
+        const updated = await userService.updateUser(user.id, update);
 
         // The update to avatarUrl is not applied
         expect(updated.avatarUrl).is.undefined;
+        expect(updated.additionalData?.disabledClosedTimeout).to.be.true;
     });
 });
